+++ activesupport/CHANGELOG.md
@@ -184,4 +184,8 @@
 
     *Daniel Schierbeck*
 
+*   Added :default option to Module delegation.
+
+    *Jack Xu*
+
 Please check [4-0-stable](https://github.com/rails/rails/blob/4-0-stable/activesupport/CHANGELOG.md) for previous changes.